Added support for union
